Change payload to Map<String,Object> so it's more versatile
This was supposed to be two separate commits, but I goofed something.

Added guard query to Operation superclass.  Modified VfModuleCreate
to store the VF count, pass it to the guard, and bump it once the
create completes successfully.
Added code to check Actors for proper plug-in to ActorService.
Renamed "operation" property to "operations", to be more consistent
with other parameters (e.g., TopicParameterGroup).
The META-INF/services files for the actors had mixed case, which
did not match the package name of the Actor class, preventing the
ServiceLoader from recognizing them.  Also modified the ActorService
to skip any that cannot actually be loaded, for whatever reason
(e.g., not in the classpath).

Issue-ID: POLICY-1625
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: Ifa97744543f2866cc553138ec5ec644b033de780