Change payload to Map<String,Object> so it's more versatile
This was supposed to be two separate commits, but I goofed something.

Added guard query to Operation superclass.  Modified VfModuleCreate
to store the VF count, pass it to the guard, and bump it once the
create completes successfully.
Added code to check Actors for proper plug-in to ActorService.
Renamed "operation" property to "operations", to be more consistent
with other parameters (e.g., TopicParameterGroup).

Issue-ID: POLICY-1625
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: Ib8c7421b035f4ef9cf08f1aa75085dd26bb5906e