Don't log cancellation exception
When an Actor operation is canceled it's done on purpose, yet the whole
exception stack trace is included in the log.  Modified the code to
leave out the stack trace for cancellations.  I don't THINK this will
cause a sonar issue, as it isn't the exception is never caught via
a "catch" clause.

Issue-ID: POLICY-1625
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: Ib5238e4791ae491b30286cdc8ed9caf67432dc26