Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> > (3 comments)
>  > 
>  > Then I think delete the legacy policy examples and provide simple
>  > examples on the latest Policy Types - folks please confirm/help me
>  > with this as this requires some work in XACML PDP.
> 
> It is OK with me to delete.   Not sure how CLAMP may be impacted.   I don't think the interface in PDP-D APPS to invoke guard changes in any way, right?

No the interface is exactly the same, but I am changing the ugly field names. Not a big change just the LHS property names, I'll submit the fixes for it.

I think for CLAMP since they never really used the old TOSCA policy types for guard, but rather sent a similar JSON as the operational policy with "content" property containing the old guard attributes, they should be fine. As long as I get the new onap.policies.guard.Common attributes to match TOSCA (which is what this JIRA is all about).

Patch-set: 2
Label: Code-Review=0