Update patch set 1
Patch Set 1:

> Does operational policy use this controller policy? I see "native"
 > embedded in this policy type name.

There is no linkage in the tosca definition between the *.drools.Controller one and the operational ones (the *Controller one derives from root same as the operational ones).   I don't see much of a relationship between operational and this controller one, other than the operational may contain a controllerName (which is optional) that may match one create with this policy.  I see the Controller one, as a native one, not operational related, which is more control loop oriented.

Patch-set: 1