Update patch set 1
Patch Set 1:

> I thought the new operational drools policy type also used
 > onap.policies.drools.Controller

In the spec, I don't see that they are linked.   I see it more a relationship with the "native.Drools" one.   The fact that the "native.Drools" one would not make sense without the "Controller" one, that's the stronger relationship, and the reason to make the controller one native.   The operational ones, don't need any of these 2 "native" policies to operate, and that's the usual case in our software.

Patch-set: 1