Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > > I thought the new operational drools policy type also used
 > >  > onap.policies.drools.Controller
 > >
 > > In the spec, I don't see that they are linked.   I see it more a
 > relationship with the "native.Drools" one.   The fact that the
 > "native.Drools" one would not make sense without the "Controller"
 > one, that's the stronger relationship, and the reason to make the
 > controller one native.   The operational ones, don't need any of
 > these 2 "native" policies to operate, and that's the usual case in
 > our software.
 > 
 > Ok - I think we were thinking there was ONE definition for a
 > controller. And that definition is shared between drools
 > operaitonal and drools native.

I see, it's not defined in the tosca schemas, there is a loose relationship based on an optional field (controllerName) that's the only commonality (but at runtime).  In any case, I really think that Drools and Controller have a strong bound (a dependence relationship, the Drools one makes no sense without the controller one, and the drools ones is plainly a rules artifact).   The relationship between the operational and any of these 2 is very loose, if any, the operational does not need any of these 2 to operate.

Patch-set: 1
Label: Code-Review=0