Add optional controllerName in Op. Legacy Policies
Optional controllerName has been added to Legacy Operational
policies to be able to route Operational Policies to an
specific drools controller.

Currently, the policies are deployed to all controllers that
support a policy type.   In the case, that there are multiple
controllers set, and the policy is only to be routed to one of
them, this field should be populated.

Issue-ID: POLICY-2388
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: I7288ce0069a723b8e6ec535331dc9cda56fbf130