Update new SDNR actor with v2.0 structures
SDNR v2.0 introduced an extra "input" and "output" layer within
the body of the messages.  Added POJOs to match that, and updated
the actor to use the new POJOs so that no extra steps are needed
during serialization or deserialization.
Also modified the SDNR and APPC-LCM simulators to discard request messages
when looking for the response, if the sink and source topics happen to
be the same.
Added the SDNR simulator.
Added more coverage to the SDNR actor.
Fixed a sonar issue in the SDNR junit code:
- use a constant instead of a literal (i.e., for "ModifyConfig")

Issue-ID: POLICY-2434
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I70914a299dd5e9ee8eaccb5d42a2596cab814091