Discard requests in APPC legacy actor
The new actor for legacy APPC attempts to discard request
messages received on the response topic by discarding those
that have a null response.  Unfortunately, when it decodes them,
the Response object creates a status object, even if there isn't
one in the json message, thus the test fails and the actor treats
it as a failed response.

Issue-ID: POLICY-2434
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I26cfb25443ff2ed394781d309f9097d50b8f160b
Signed-off-by: Jim Hahn <jrh3@att.com>