Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> When I tried upgrading lombok, I saw a lot of this. Wasn't sure about it and didn't have enough time to investigate it. If someone has cycles, they may want to upgrade parent to the lates lombok and see if these can get cleared out.

I looked at this briefly a few months ago. What's happening is that the Lombok version used in Eclipse is not the same as that used in the build and the exception message has changed. I was afraid to touch it in Parent in case it broke more unit tests that expect the old message test.

Patch-set: 2
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>