Update patch set 1
Patch Set 1: Code-Review-1

Liam, is it possible to create an independent bug for the concurrency issues, it is pointing to an epic now?   I think the issue may be noticeable by other folks in labs, and want to point them out to a ticket that tracks it.   I think this may be the issue that Brian is running into and want to give it a ticket.   I also agree with the comments from Jim.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1