Cannot parse finishTime in SO responses
The timestamp in SO responses appears to follow RFC-1123/RFC-822
format instead of ISO format.
Added a type adapter for SO.  In Guilin, we may want to consider
combine the two type adapters for LocalDateTime so that either format
is accepted.
Also modified the SO simulator to return responses that are actual
samples from SO.
As part of that work, discovered that the legacy SO actor is not able
to parse these timestamps either.  However, as that code is now
deprecated, left it as is.

Issue-ID: POLICY-2570
Change-Id: I322318d1007e36eef47bb8867fd8ed01cb60223a
Signed-off-by: Jim Hahn <jrh3@att.com>