Cannot parse finishTime in legacy SO responses
Adding the actual sample responses to the SO simulator broke the
drools-apps junit for usecases.  Fixed (in theory) by updating the
legacy SO actor to properly decode the finishTime.  Refactored the
new SO actor, extracting the type adapter into its own class file
so it could be shared between the new and legacy actors.

Issue-ID: POLICY-2570
Change-Id: I061b603172440b1a91da16d09b4f2a0d289dfc41
Signed-off-by: Jim Hahn <jrh3@att.com>