Copy Dbao from drools-apps to models
Drools-apps and xacml-pdp each have their own copy of Dbao.  Made
a copy of it in models as a first step toward eliminating the other
copies.  Also changed its package name be consistent with the other
interface package names and renamed it to OperationsHistory.
Also reordered the modules in the pom to make it alphabetical.

Issue-ID: POLICY-2581
Change-Id: I45afa680c6344eeac6b2d35c9e4f04a5ba7af431
Signed-off-by: Jim Hahn <jrh3@att.com>