Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> should probably fix it in the yaml, too.  I'm not sure of the difference between the input.tosca and the output.tosca files, but may need to fix it in the output.tosca files, as well

I understand that in yaml the quoting is optional,that's the reason I didn't do changes.   Other fields don't have it.

For the purpose of this ticket to fix the csit issue, as making strings explicit in the yaml, will impact most of the fields in all policies.

Another take would be to quote only the modelVersion for the reader to make it explicit, but not sure if worth it, as the tosca policy type schema should make it clear that is a string.

But I am open to your suggestions, let me know.

Patch-set: 1