Fix broken models build due to json upgrade
The json dependency was moved to policy-parent and the "version"
property removed which broke the policy-models build.  Removed
the "version" from the pom.
Also removed the jetty-security "version" from the pom, as it is
no longer needed either.
Also added asm dependency, because upgrade of openpojo version
caused breakage in one of the TestPojos.

Issue-ID: POLICY-2387
Change-Id: I3c2315b3cbea05de76b6dfcc289f9323af77745b
Signed-off-by: Jim Hahn <jrh3@att.com>