Add property lists to Actors
Modified the Actor code to provide a list of properties needed to by the
actor to perform a given operation.
Added a build() method to the parameter class so invokers can build an
operation and set its properties prior to starting it.
Added a "preprocessed" field to the parameter class so invokers can
indicate that the Actor need not perform any preprocessing steps.  Will
modify the actors, in a subsequent review, to observe the flag.
Added "properties" to Operation so invokers can set the properties.

Issue-ID: POLICY-2746
Change-Id: I65996aef5cec5afe25e8287c0b2f5f322c532ca5
Signed-off-by: Jim Hahn <jrh3@att.com>