Remove junit annotations from dmaap common test class
Jenkins builds are randomly failing in the junits for the dmaap simulator.
From the console output, it appears that simulator is being shut down
before all of the test methods have executed.  Moved the code for
managing the "Main" object from the common superclass to the one
subclass that needed it.

Issue-ID: POLICY-2749
Change-Id: I2615bbef9b790dbb633ed7f123aeb3f880826862
Signed-off-by: Jim Hahn <jrh3@att.com>