Move property names to Operator
When an Operation is constructed, it assumes that the ControlLoopParams
is fully populated.  However, the data may not all be available until
getPropertyNames() has been called and the preprocessing steps executed.
This creates a catch-22 situation.  Solved the problem by moving the
getPropertyNames() method from the Operation class to the Operator class.
Also fixed a case in A&AI post-processing steps to prevent NPE when
the event context is not provided.

Issue-ID: POLICY-2746
Change-Id: I2595ba83f181c6028b93ce6817744b9c37eb44c5
Signed-off-by: Jim Hahn <jrh3@att.com>