Default should be an object on TOSCA properties
Note that in this implementation, when the object is re-read from the
database, the default value will be a string representation of the
original default value, not the actual original default value.

In order to do correct type conversion from the string stored in the
database to the stored type, the Policy Framework TOSCA handling code
would have to have code that understands and converts the types in the
TOSCA itself. In other words, it would have to understand what the TOSCA
type is and understand how to serialize and deserialize it.

Issue-ID: POLICY-2844
Change-Id: I69f2843a93cd5f831359a0965223298b769498d9
Signed-off-by: liamfallon <liam.fallon@est.tech>