Make simulator parameters optional
Modified the simulator to make the DMaaP and GRPC parameters optional so
that a simulator can be started without them.  Also added parameter
validation, as that was not being performed on the DMaaP or GRPC
parameters.  Also modified the validation so-as not to require the rest
server parameter field contained within the DMaaP parameter class.

Issue-ID: POLICY-2858
Change-Id: I85dde93c7baf863eca2cb10eab37806f3e08a087
Signed-off-by: Jim Hahn <jrh3@att.com>