Delete preprocessed flag from actors
Removed the "preprocessed" flag from the Actor parameter class, now
that the actors no longer have a startPreprocess() method.

Also removed targetEntity from Actor parameter class.  Created a
makeOutcome method within OperationPartial, which is used to create an
initial outcome whose target field is pre-populated with the
target-entity extracted from the properties.  As the meaning of "target"
may be specific to an operation, the makeOutcome method may be
overridden by an operation subclass.

Issue-ID: POLICY-2804
Change-Id: Ifb66de63301d644e69340009593513773ee5672d
Signed-off-by: Jim Hahn <jrh3@att.com>