Work around docker rate limit in CSITs
CSITs are failing regularly due to download rate limits that were
recently introduced on dockerhub.  Modified the Dockerfiles for the
simulators that are used by the CSITs to pull the base images from
nexus3 instead of dockerhub.
Based on feedback from Steve Winslow, it's OK to pull base images from
nexus3 provided that the docker images generated from it are not
actually released.  In the case of the simulators, only snapshot images
are ever created.
Also added the policy alpine version number to ensure the appropriate
docker base is used for frankfurt.

Issue-ID: POLICY-2895
Change-Id: I61393f3551c2257b75948f39add0b5163265e152
Signed-off-by: Jim Hahn <jrh3@att.com>