Update patch set 1
Patch Set 1:

Right now, no deliverables depend on this version of models, but if we do need a change to models for Honolulu, then do we want this review included, too?  Or should we branch models so this is only in Istanbul?

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>