Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> Hi Jim, the control loop test cases works fine with timestamp which are using PfReferenceTimestamp and PfTimestamp keys. Not seeing any issues with db. I am not sure I understand, why are we facing this only on pap.

Appears to be one of those things that causes different results depending on the environment.  So without this change, it fails when the PAP junit is run in my eclipse, it works in the jenkins build, and it fails when PAP is run in kubernetes.  I don't think it effects the other repos as the classes they use don't have timestamp keys.

Obviously, we want to avoid code that results in failures in one environment but not another.

Just guessing, but I suspect it's related to the fact that java.sql.Timestamp has two different internal structures for tracking the time, only one of which need contain data.  If the one containing "nanos" is used, but the one retrieved from the DB uses the "date" structure instead, then there's a mismatch when the EntityManager.refresh() method is invoked - that's where the junit is failing, saying it can't find the record that was just stored/retrieved.

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>