Update patch set 2
Patch Set 2:

> Is it possible to have an option to disable them system wide on the PDP-D instead of a per policy basis, similar to the guard.disabled, or perhaps by removing the guard actor configuration?

the desire was to be able to disable guards on a per-policy basis, so disabling it system wide doesn't address the issue.  However, one option, that would not require any code change, would be to simply segregate PDP-Ds into two different subgroups so that some are configured with guard.disabled=true, while the others are configured with guard.disabled=false.

Another option would be to introduce the guardDisabled flag in version 1.2.0 of the common.Drools policy type, so it doesn't impact existing policies and policy types.  Then in PDP-D, we could introduce a second json schema to validate against that version, or we could just use the updated json schema in the associated review, as I believe it will work with policies of either type (although it would allow policies of type 1.1.0 to include a guardDisabled flag).

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>