Update patch set 2
Patch Set 2:

Just a thought: why not add a 'guardDisabled' flag under the operational policy's operations payload? Each operation in the policy chain would specify whether to invoke guard or not by specifying 'guardDisabled' in the payload and thus satisfies the desire to have guard enabled/disabled on a per policy basis and does not require a model change. It would need a change in the Guard actor to check the payload's 'guardDisabled' flag over any configured environment value I think.

Patch-set: 2