Update patch set 2
Patch Set 2:

in the future, I would imagine that a guard check would just be one more operation in the list of operations.  Once that approach is taken, there would no longer be a need for flags at the policy or operation level.  OTOH, there would still need to be a way for PDP-D to know whether or not guards were embedded within the list of operations - maybe use another policy type (which just happens to look like the existing policy type).  We could accelerate the implementation of this approach to make it happen in Istanbul.

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>