Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hmm.  I'm questioning the need for POLICY-3256, when deleteByConceptKey() can already be used to delete particular templates by name & version - seems safer than deleting by name only.

Yes, we need to think about this more. The deelteByConceptkey() is on Dao level but is not exposed above that. Also what worries me is referential integrity, of a concept is referred to by another concept then should we allow it to be deleted? Also though, maybe we need a "super user" ability so that a "super user" can update and delete concepts in the unlikely (;-D) event of things getting mucked up.

Also I'd like to have a more generic way of doing CRUD for TOSCA concepts, we have a good few TOSCA concepts int here now that are all almost identical (type and corresponding instance), if we have a bespoke API that covers all the concepts it's going to get very unwieldy

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>