No-sonar on the wrong line
Tried putting NOSONAR comment at the end of the statement, because it
seems like checkstyle/sonar doesn't like comments in the middle of a
statement.  Nonetheless, sonar still reports an issue on the line with
the "SELECT" statement, so I guess there's no choice but to move NOSONAR
to the end of that line.

Issue-ID: POLICY-3094
Change-Id: If26ad0b1e7dcc9c7aa963da10a9c5160a8fe272d
Signed-off-by: Jim Hahn <jrh3@att.com>