Update patch set 3
Patch Set 3:

(1 comment)

Do you mind to summarize for me the outcome of the thread conversation, I couldn't quite capture what the resolution is going to be?   
In particular defining what the following fields are, and how they are going to be constructed:
- PdpMessage.name   #  currently this is the hostname PDPs -> PAP (not sure in the other dir PAP->PDPs)
- [PdpUpdate|PdpStateChange].source   # hostname + id ??
It would be good too to add it too in the code in the classes as javadoc to the data member field for reference as I will forget :).

Patch-set: 3
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>