Delete extra simulator docker images
The policy-models-simulator docker image, built from
models-sim/packages, subsumes the functionality of the other simulator
images.  Deleted the other images.

This simulator still needs some of the code from the dmaap-sim modules,
so only deleted the code from the latter that is not needed by this
simulator.

Also modified the simulator launch script to use sh instead of bash.

Issue-ID: POLICY-3474
Change-Id: Ie7f69b327a18c3e07f154124ae38975d5e6924f0
Signed-off-by: Jim Hahn <jrh3@att.com>