Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-2
> > > 
> > > this simulator should not be used; it's functionality is entirely subsumed by policy-models-simulator
> > 
> > OK, let me check that....
> 
> you can find a sample config file in policy-docker/csit/config/sim-all.  If just using it for dmaap simulation, all you need is the dmaapProvider section and the "DMaaP simulator" portion of the restServers section.  Everything else can be deleted (it's possible that it needs the topicSinks/Sources/Servers sections - if so, just provide empty arrays, [], for those).

what we could do, though, is add a default set of simParameters.json to the docker image, that just has what's needed for the dmaap portion, and add a flag to simulators.sh so that it can be launched without specifying any keystore/truststore info.  That would make it easier to use as a simulator, as-is.

Patch-set: 1