Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> @Jorge I checked the drools CSITs and they seem to be working fine with this change. Is there any other place I should check where drools is using the AAI simulator?

The aai simulator is used in the feature-controlloop-utils in drools-applications.   The s3p.jmx for jmeter in drools-applications is used for vfirewall, vcpe, .. and in the ONSETs carry the AAI info for the queries, it is typically used with the feature-controlloop-utils  The drools-applications junits for the usecases controller also run the queries against the simulator.

Patch-set: 5
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>