Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > @Jorge I checked the drools CSITs and they seem to be working fine with this change. Is there any other place I should check where drools is using the AAI simulator?
> 
> The aai simulator is used in the feature-controlloop-utils in drools-applications.   The s3p.jmx for jmeter in drools-applications is used for vfirewall, vcpe, .. and in the ONSETs carry the AAI info for the queries, it is typically used with the feature-controlloop-utils  The drools-applications junits for the usecases controller also run the queries against the simulator.

The only change in behavior to the existing AAI simulator is with GetPnfQuery. This needs correct pnf name to fetch the corresponding response file. Sample response which was already there can be fetched by using "demo-pnf" as the pnf name. Probably no changes will be needed anywhere similar to CSITs.

Patch-set: 5