Update patch set 5
Patch Set 5: Code-Review+1

drools-applications is also verified and the junits seems to pass okay.
I don't think there will be any changes needed in jmx as well. Even if it is needed, it will be as simple as updating the pnf-name, as this is an improvement and provide more flexibility anyway.
So, I think we are good.

Patch-set: 5
Reviewer: Gerrit User 4903 <4903@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1